requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask-migrate==4.0.4
requests==2.31.0
google-generativeai==0.3.1


pypdf==3.15.1
pdfplumber==0.10.2
python-docx==0.8.11
odfpy==1.4.1